Disable write batching for PSQL
Problem: batches of transactions are accumulated before an EOF of trace
file is encountered. Given files are written in parallel by the fetcher,
this may lead to potentially indefinite accumulation of batches without
commitment.

Solution: disable batching for Postgresql (as it was originally
introduced as a necessary optimization for sqlite).